backend/requirements.txt
coverage==6.4.4
Django==4.1.1
psycopg2-binary==2.9.3
gunicorn==20.1.0
djangorestframework==3.13.1
django-ckeditor==6.5.1
drf-yasg==1.21.3
factory-boy==3.2.1
Faker==15.0.0
Pillow==9.2.0
six==1.16.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
PyYAML==6.0
ytmusicapi==0.22.0